# packages/cactus-plugin-ledger-connector-fabric/src/main/json/openapi.json
{
    "openapi": "3.0.3",
    "info": {
        "title": "Hyperledger Cactus Plugin - Connector Fabric",
        "description": "Can perform basic tasks on a fabric ledger",
        "version": "0.0.1",
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        }
    },
    "servers": [
        {
            "url": "https://www.cactus.stream/{basePath}",
            "description": "Public test instance",
            "variables": {
                "basePath": {
                    "default": ""
                }
            }
        },
        {
            "url": "http://localhost:4000/{basePath}",
            "description": "Local test instance",
            "variables": {
                "basePath": {
                    "default": ""
                }
            }
        }
    ],
    "components": {
        "schemas": {
            "FabricSigningCredential": {
                "type": "object",
                "required": [
                    "type"
                ],
                "discriminator": {
                    "propertyName": "type"
                },
                "oneOf": [
                    {
                        "$ref": "#/components/schemas/FabricSigningCredentialCactusKeychainRef"
                    },
                    {
                        "$ref": "#/components/schemas/FabricSigningCredentialNone"
                    }
                ],
                "properties": {
                    "type": {
                        "$ref": "#/components/schemas/FabricSigningCredentialType"
                    }
                }
            },
            "FabricSigningCredentialCactusKeychainRef": {
                "type": "object",
                "required": [
                    "keychainId",
                    "keychainRef",
                    "type"
                ],
                "properties": {
                    "keychainId": {
                        "type": "string",
                        "minLength": 1,
                        "maxLength": 100,
                        "nullable": false
                    },
                    "keychainRef": {
                        "type": "string",
                        "minLength": 1,
                        "maxLength": 100,
                        "nullable": false
                    },
                        "type": {
                            "$ref": "#/components/schemas/FabricSigningCredentialType"
                        }
                    }
                },
            "FabricSigningCredentialNone": {
                "type": "object",
                "required": [
                    "type"
                ],
                "description": "Using this denotes that there is no signing required because the transaction is pre-signed.",
                "properties": {
                    "type": {
                        "$ref": "#/components/schemas/FabricSigningCredentialType"
                    }
                }
            },
            "FabricSigningCredentialType": {
                "type": "string",
                "enum": [
                    "CACTUS_KEYCHAIN_REF",
                    "NONE"
                ]
            },
...
            "UnsignedProposalRequest": {
                "type": "object",
                "required": [
                    "channelName",
                    "chaincodeName",
                    "functionName",
                    "functionArgs",
                    "signingCredential",
                    "serviceUserName"
                ],
                "properties": {
                    "channelName": {
                        "type": "string",
                        "nullable": false
                    },
                    "chaincodeName": {
                        "type": "string",
                        "nullable": false
                    },
                    "functionName": {
                        "type": "string",
                        "nullable": false
                    },
                    "functionArgs": {
                        "type": "array",
                        "nullable": false,
                        "default": [],
                        "items": {
                            "type": "string",
                            "nullable": true
                        }
                    },
                    "signingCredential": {
                        "$ref": "#/components/schemas/FabricSigningCredential",
                        "nullable": false
                    },
                    "serviceUserIdentity": {
                        "type": "string"
                    },
                    "serviceUserName": {
                        "type": "string",
                        "default": "",
                        "nullable": false
                    }
                }
            },
            "UnsignedProposalResponse": {
                "type": "object"
            },
            "SendSignedEndorsementRequest": {
                "type": "object",
                "required": [
                    "endorsement",
                    "signedProposal",
                    "serviceUserName",
                    "channelName",
                    "organisation"
                ],
                "properties": {
                    "endorsement": {
                        "type": "object",
                        "nullable": false
                    },
                    "signedProposal": {
                        "type": "object",
                        "nullable": false
                    },
                    "serviceUserIdentity": {
                        "type": "string"
                    },
                    "signingCredential": {
                        "$ref": "#/components/schemas/FabricSigningCredential",
                        "nullable": false
                    },
                    "serviceUserName": {
                        "type": "string",
                        "default": "",
                        "nullable": false
                    },
                    "channelName": {
                        "type": "string",
                        "nullable": false
                    },
                    "organisation": {
                        "type": "string",
                        "nullable": false
                    }
                }
            },
            "SendSignedEndorsementResponse": {
                "type": "object"
            },
...
        }
    },
    "paths": {
...
        "/api/v1/plugins/@hyperledger/cactus-plugin-ledger-connector-fabric/send-signed-endorsement": {
            "post": {
                "x-hyperledger-cactus": {
                    "http": {
                        "verbLowerCase": "post",
                        "path": "/api/v1/plugins/@hyperledger/cactus-plugin-ledger-connector-fabric/send-signed-endorsement"
                    }
                },
                "operationId": "sendSignedEndorsement",
                "summary": "Sends a pre-signed endorsement to the Fabric ledger.",
                "description": "",
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/SendSignedEndorsementRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SendSignedEndorsementResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": ""
                    }
                }
            }
        },
        "/api/v1/plugins/@hyperledger/cactus-plugin-ledger-connector-fabric/generate-unsigned-proposal": {
            "post": {
                "x-hyperledger-cactus": {
                    "http": {
                        "verbLowerCase": "post",
                        "path": "/api/v1/plugins/@hyperledger/cactus-plugin-ledger-connector-fabric/generate-unsigned-proposal"
                    }
                },
                "operationId": "generateUnsignedProposal",
                "summary": "Sends a pre-signed proposal to the Fabric ledger.",
                "description": "",
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UnsignedProposalRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UnsignedProposalResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": ""
                    }
                }
            }
        },
...
    }
}